Cut data to a maximal given time

Cut data to a maximal given time

cut_data(data, Tmax)

Arguments

data

data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

Tmax

max time considered

Value

a data.frame with the same format as data where each individual has Tmax as last time entry.

Author

Cristian Preda

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement
K <- 4
PJK <- matrix(1 / 3, nrow = K, ncol = K) - diag(rep(1 / 3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)
tail(d_JK)
#>     id     time state
#> 103 10 5.745181     2
#> 104 10 6.399224     3
#> 105 10 6.658211     2
#> 106 10 7.359727     1
#> 107 10 8.174763     4
#> 108 10 9.045728     1

# cut at Tmax = 8
d_JK2 <- cut_data(d_JK, Tmax = 8)
tail(d_JK2)
#>    id     time state
#> 93 10 4.895024     4
#> 94 10 5.745181     2
#> 95 10 6.399224     3
#> 96 10 6.658211     2
#> 97 10 7.359727     1
#> 98 10 8.000000     1