Estimate probabilities to be in each state

Estimate probabilities to be in each state

estimate_pt(data, NAafterTmax = FALSE)

Arguments

data
data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

NAafterTmax
if TRUE, return NA if t > Tmax otherwise return the state associated with Tmax (useful when individuals has different lengths)

Value

A list of two elements:

t: vector of time

pt: a matrix with K (= number of states) rows and with length(t) columns containing the probabilities to be in each state at each time.

See also

Author

Cristian Preda, Quentin Grimonprez

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement 
K <- 4
PJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)

d_JK2 <- cut_data(d_JK, 10)

# estimate probabilities
estimate_pt(d_JK2)

#> $pt
#>   0 0.218 1.06 1.082 1.225 1.364 1.651 1.758 1.797 1.86 1.887 1.898 2.315 2.36
#> 1 1   0.9  0.8   0.8   0.7   0.6   0.5   0.4   0.4  0.3   0.2   0.2   0.1  0.0
#> 2 0   0.0  0.1   0.1   0.1   0.2   0.2   0.3   0.4  0.4   0.5   0.4   0.4  0.4
#> 3 0   0.1  0.1   0.0   0.1   0.1   0.1   0.1   0.1  0.1   0.1   0.1   0.1  0.2
#> 4 0   0.0  0.0   0.1   0.1   0.1   0.2   0.2   0.1  0.2   0.2   0.3   0.4  0.4
#>   2.396 2.414 2.416 2.452 2.456 2.876 2.949 3.034 3.136 3.158 3.228 3.228 3.353
#> 1   0.0   0.0   0.0   0.1   0.1   0.0   0.1   0.1   0.0   0.0   0.0   0.0   0.0
#> 2   0.3   0.4   0.3   0.3   0.3   0.3   0.3   0.4   0.4   0.5   0.4   0.5   0.6
#> 3   0.2   0.2   0.3   0.2   0.1   0.1   0.0   0.0   0.0   0.0   0.0   0.0   0.0
#> 4   0.5   0.4   0.4   0.4   0.5   0.6   0.6   0.5   0.6   0.5   0.6   0.5   0.4
#>   3.387 3.478 3.481 3.579 3.615 3.659 3.707 3.731 3.735 3.754 3.767 3.842 3.865
#> 1   0.1   0.1   0.1   0.1   0.1   0.0   0.0   0.0   0.1   0.1   0.1   0.1   0.1
#> 2   0.5   0.6   0.5   0.5   0.4   0.4   0.3   0.3   0.3   0.4   0.3   0.2   0.2
#> 3   0.0   0.0   0.1   0.2   0.2   0.3   0.4   0.5   0.4   0.3   0.4   0.5   0.4
#> 4   0.4   0.3   0.3   0.2   0.3   0.3   0.3   0.2   0.2   0.2   0.2   0.2   0.3
#>   3.933 4.178 4.205 4.411 4.459 4.739 4.799 4.843 4.86 4.868 4.947 5.108 5.161
#> 1   0.1   0.2   0.3   0.2   0.2   0.1   0.0   0.0  0.0   0.0   0.1   0.2   0.2
#> 2   0.3   0.2   0.2   0.2   0.1   0.1   0.1   0.1  0.2   0.3   0.2   0.1   0.2
#> 3   0.3   0.3   0.3   0.3   0.3   0.3   0.3   0.4  0.3   0.3   0.3   0.3   0.2
#> 4   0.3   0.3   0.2   0.3   0.4   0.5   0.6   0.5  0.5   0.4   0.4   0.4   0.4
#>   5.257 5.27 5.372 5.399 5.499 5.596 5.658 5.702 5.744 5.814 5.899 5.924 5.947
#> 1   0.1  0.1   0.1   0.1   0.1   0.1   0.1   0.0   0.1   0.1   0.1   0.1   0.0
#> 2   0.2  0.3   0.4   0.5   0.4   0.4   0.5   0.6   0.6   0.5   0.4   0.3   0.3
#> 3   0.3  0.2   0.1   0.1   0.1   0.2   0.2   0.2   0.1   0.2   0.3   0.4   0.4
#> 4   0.4  0.4   0.4   0.3   0.4   0.3   0.2   0.2   0.2   0.2   0.2   0.2   0.3
#>   5.949 6.033 6.034 6.037 6.135 6.157 6.176 6.19 6.192 6.663 6.712 6.845 6.944
#> 1   0.1   0.1   0.2   0.3   0.2   0.2   0.1  0.1   0.1   0.1   0.1   0.1   0.0
#> 2   0.3   0.2   0.2   0.2   0.3   0.4   0.5  0.4   0.4   0.3   0.3   0.2   0.2
#> 3   0.4   0.5   0.4   0.3   0.3   0.3   0.3  0.3   0.2   0.3   0.4   0.4   0.5
#> 4   0.2   0.2   0.2   0.2   0.2   0.1   0.1  0.2   0.3   0.3   0.2   0.3   0.3
#>   7.002 7.054 7.091 7.231 7.365 7.39 7.633 7.698 7.884 7.959 8.153 8.267 8.435
#> 1   0.0   0.0   0.0   0.0   0.0  0.0   0.0   0.1   0.1   0.1   0.1   0.2   0.2
#> 2   0.1   0.1   0.2   0.2   0.2  0.3   0.3   0.3   0.4   0.5   0.6   0.5   0.4
#> 3   0.6   0.5   0.4   0.5   0.6  0.6   0.5   0.4   0.3   0.3   0.2   0.2   0.3
#> 4   0.3   0.4   0.4   0.3   0.2  0.1   0.2   0.2   0.2   0.1   0.1   0.1   0.1
#>   8.448 8.5 8.509 8.933 8.969 9.007 9.118 9.17 9.175 9.193 9.235 9.505 9.509
#> 1   0.2 0.2   0.2   0.1   0.2   0.2   0.2  0.1   0.1   0.0   0.0   0.0   0.1
#> 2   0.5 0.5   0.5   0.6   0.6   0.5   0.4  0.5   0.4   0.5   0.4   0.5   0.5
#> 3   0.2 0.1   0.0   0.0   0.0   0.1   0.1  0.1   0.1   0.1   0.1   0.1   0.1
#> 4   0.1 0.2   0.3   0.3   0.2   0.2   0.3  0.3   0.4   0.4   0.5   0.4   0.3
#>   9.522 9.616 9.695 9.846  10
#> 1   0.1   0.1   0.1   0.0 0.0
#> 2   0.6   0.5   0.6   0.6 0.6
#> 3   0.1   0.1   0.1   0.2 0.2
#> 4   0.2   0.3   0.2   0.2 0.2
#> 
#> $t
#>   [1]  0.000000  0.218029  1.059869  1.082104  1.225225  1.364102  1.650565
#>   [8]  1.758437  1.796691  1.860432  1.886523  1.898256  2.314744  2.359585
#>  [15]  2.396286  2.414224  2.416448  2.451532  2.455921  2.875623  2.949086
#>  [22]  3.034474  3.136485  3.158225  3.227637  3.228351  3.352949  3.387068
#>  [29]  3.477895  3.481108  3.579120  3.614924  3.659451  3.706632  3.730527
#>  [36]  3.735177  3.753774  3.766645  3.841960  3.864564  3.933192  4.178203
#>  [43]  4.205289  4.411318  4.459122  4.739096  4.799184  4.842965  4.860134
#>  [50]  4.867784  4.947494  5.107560  5.160510  5.257211  5.270245  5.371948
#>  [57]  5.399494  5.498711  5.596431  5.658369  5.701601  5.743582  5.814145
#>  [64]  5.898758  5.924317  5.947263  5.948600  6.032843  6.034007  6.037363
#>  [71]  6.135105  6.156928  6.176082  6.189515  6.192005  6.662948  6.712156
#>  [78]  6.845386  6.943925  7.001735  7.054134  7.090929  7.231146  7.365274
#>  [85]  7.389744  7.632948  7.697710  7.883571  7.958644  8.153078  8.266830
#>  [92]  8.435319  8.447659  8.500367  8.509131  8.932791  8.969060  9.006710
#>  [99]  9.117689  9.170231  9.174840  9.193178  9.235290  9.505369  9.508954
#> [106]  9.521550  9.615911  9.695191  9.846419 10.000000
#> 
#> attr(,"class")
#> [1] "pt"