Table of transitions

Calculates a frequency table counting the number of times each pair of states were observed in successive observation times.

statetable(data, removeDiagonal = FALSE)

Arguments

data
data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

removeDiagonal
if TRUE, does not count transition from a state i to i

Value

a vector of length K containing the total time spent in each state

Author

Quentin Grimonprez

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement 
K <- 4
PJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)

# table of transitions
statetable(d_JK)

#>     to
#> from  1  2  3  4
#>    1  0 14 10 11
#>    2 10  0  7  4
#>    3  8  6  0  9
#>    4 11  5  6  0