Summary

Get a summary of the data.frame containing categorical functional data

summary_cfd(data, max.print = 10)

Arguments

data

data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

max.print

maximal number of states to display

Value

a list containing:

nRow number of rows

nInd number of individuals

timeRange minimal and maximal time value

uniqueStart TRUE, if all individuals have the same time start value

uniqueEnd TRUE, if all individuals have the same time start value

states vector containing the different states

visit number of individuals visiting each state

See also

Author

Quentin Grimonprez

Examples

data(biofam2)
summary_cfd(biofam2)
#> Number of rows: 6965 
#> Number of individuals: 2000 
#> Time Range: 15 - 30 
#> Same time start value for all ids: TRUE 
#> Same time end value for all ids: TRUE 
#> Number of states: 8 
#> States:
#>   Parent, Left+Marr, Left+Marr+Child, Left, Married, Divorced, Left+Child, Child
#> Number of individuals visiting each state:
#>          Parent       Left+Marr Left+Marr+Child            Left         Married 
#>            1972             987             907             896             246 
#>        Divorced      Left+Child           Child 
#>              76              40               6