Estimate transition matrix and spent time

Calculates crude initial values for transition intensities by assuming that the data represent the exact transition times of the Markov process.

estimate_Markov(data)

Arguments

data
data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

Value

list of two elements: Q, the estimated transition matrix, and lambda, the estimated time spent in each state

See also

Author

Cristian Preda

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement
K <- 4
PJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 100, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)

# estimation
mark <- estimate_Markov(d_JK)
mark$P
#>     to
#> from         1         2         3         4
#>    1 0.0000000 0.3289037 0.2956811 0.3754153
#>    2 0.3069767 0.0000000 0.3255814 0.3674419
#>    3 0.3857143 0.3190476 0.0000000 0.2952381
#>    4 0.3491379 0.3103448 0.3405172 0.0000000
mark$lambda
#>         1         2         3         4 
#> 0.9933151 1.0508401 1.1517039 1.1413771