Estimate probabilities to be in each state

Estimate probabilities to be in each state

estimate_pt(data, NAafterTmax = FALSE)

Arguments

data
data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

NAafterTmax
if TRUE, return NA if t > Tmax otherwise return the state associated with Tmax (useful when individuals has different lengths)

Value

A list of two elements:

t: vector of time

pt: a matrix with K (= number of states) rows and with length(t) columns containing the probabilities to be in each state at each time.

See also

Author

Cristian Preda, Quentin Grimonprez

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement 
K <- 4
PJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)

d_JK2 <- cut_data(d_JK, 10)

# estimate probabilities
estimate_pt(d_JK2)

#> $pt
#>   0 0.218 0.492 0.87 1.06 1.063 1.082 1.139 1.225 1.364 1.651 1.797 1.887 1.898
#> 1 1   0.9   0.8  0.7  0.6   0.7   0.7   0.6   0.5   0.4   0.3   0.3   0.2   0.2
#> 2 0   0.0   0.1  0.1  0.2   0.2   0.2   0.3   0.3   0.4   0.4   0.5   0.6   0.5
#> 3 0   0.1   0.1  0.1  0.1   0.1   0.0   0.0   0.1   0.1   0.1   0.1   0.1   0.1
#> 4 0   0.0   0.0  0.1  0.1   0.0   0.1   0.1   0.1   0.1   0.2   0.1   0.1   0.2
#>   2.315 2.324 2.36 2.396 2.414 2.416 2.452 2.456 2.509 2.876 2.949 3.034 3.136
#> 1   0.1   0.2  0.1   0.1   0.1   0.1   0.2   0.2   0.2   0.1   0.2   0.2   0.1
#> 2   0.5   0.4  0.4   0.3   0.4   0.3   0.3   0.3   0.2   0.2   0.2   0.3   0.3
#> 3   0.1   0.1  0.2   0.2   0.2   0.3   0.2   0.1   0.2   0.2   0.1   0.1   0.1
#> 4   0.3   0.3  0.3   0.4   0.3   0.3   0.3   0.4   0.4   0.5   0.5   0.4   0.5
#>   3.158 3.228 3.228 3.353 3.387 3.481 3.579 3.659 3.707 3.731 3.735 3.754 3.767
#> 1   0.1   0.1   0.1   0.1   0.2   0.2   0.2   0.1   0.1   0.1   0.2   0.2   0.2
#> 2   0.4   0.3   0.4   0.5   0.4   0.3   0.3   0.3   0.2   0.2   0.2   0.3   0.2
#> 3   0.1   0.1   0.1   0.1   0.1   0.2   0.3   0.4   0.5   0.6   0.5   0.4   0.5
#> 4   0.4   0.5   0.4   0.3   0.3   0.3   0.2   0.2   0.2   0.1   0.1   0.1   0.1
#>   3.842 3.865 3.933 4.061 4.205 4.411 4.459 4.486 4.739 4.843 4.86 4.868 4.947
#> 1   0.2   0.2   0.2   0.1   0.2   0.1   0.1   0.2   0.1   0.1  0.1   0.1   0.2
#> 2   0.1   0.1   0.2   0.2   0.2   0.2   0.1   0.1   0.1   0.1  0.2   0.3   0.2
#> 3   0.6   0.5   0.4   0.5   0.5   0.5   0.5   0.4   0.4   0.5  0.4   0.4   0.4
#> 4   0.1   0.2   0.2   0.2   0.1   0.2   0.3   0.3   0.4   0.3  0.3   0.2   0.2
#>   4.966 5.019 5.108 5.161 5.257 5.27 5.372 5.497 5.499 5.596 5.658 5.702 5.744
#> 1   0.1   0.1   0.2   0.2   0.1  0.1   0.1   0.1   0.1   0.1   0.1   0.0   0.1
#> 2   0.3   0.2   0.1   0.2   0.2  0.3   0.4   0.4   0.3   0.3   0.4   0.5   0.5
#> 3   0.4   0.4   0.4   0.3   0.4  0.3   0.2   0.1   0.1   0.2   0.2   0.2   0.1
#> 4   0.2   0.3   0.3   0.3   0.3  0.3   0.3   0.4   0.5   0.4   0.3   0.3   0.3
#>   5.814 5.899 5.924 5.928 5.947 5.949 6.034 6.135 6.192 6.4 6.662 6.845 6.859
#> 1   0.1   0.1   0.1   0.2   0.1   0.2   0.3   0.2   0.2 0.1   0.1   0.1   0.2
#> 2   0.4   0.3   0.2   0.2   0.2   0.2   0.2   0.3   0.3 0.3   0.4   0.3   0.3
#> 3   0.2   0.3   0.4   0.4   0.4   0.4   0.3   0.3   0.2 0.2   0.2   0.2   0.2
#> 4   0.3   0.3   0.3   0.2   0.3   0.2   0.2   0.2   0.3 0.4   0.3   0.4   0.3
#>   6.944 7.002 7.054 7.231 7.365 7.39 7.633 7.698 7.884 7.959 8.091 8.175 8.435
#> 1   0.1   0.1   0.1   0.1   0.1  0.1   0.1   0.2   0.2   0.2   0.1   0.1   0.1
#> 2   0.3   0.2   0.2   0.2   0.2  0.3   0.3   0.3   0.4   0.5   0.5   0.4   0.3
#> 3   0.3   0.4   0.3   0.4   0.5  0.5   0.4   0.3   0.2   0.2   0.3   0.3   0.4
#> 4   0.3   0.3   0.4   0.3   0.2  0.1   0.2   0.2   0.2   0.1   0.1   0.2   0.2
#>   8.448 8.5 8.509 8.78 8.825 8.933 8.969 9.007 9.118 9.193 9.495 9.505 9.616
#> 1   0.1 0.1   0.1  0.2   0.3   0.2   0.3   0.3   0.3   0.2   0.1   0.1   0.1
#> 2   0.4 0.4   0.4  0.4   0.4   0.5   0.5   0.4   0.3   0.4   0.4   0.5   0.4
#> 3   0.3 0.2   0.1  0.1   0.0   0.0   0.0   0.1   0.1   0.1   0.1   0.1   0.1
#> 4   0.2 0.3   0.4  0.3   0.3   0.3   0.2   0.2   0.3   0.3   0.4   0.3   0.4
#>   9.695  10
#> 1   0.1 0.1
#> 2   0.5 0.5
#> 3   0.1 0.1
#> 4   0.3 0.3
#> 
#> $t
#>   [1]  0.0000000  0.2180290  0.4916986  0.8704080  1.0598693  1.0630830
#>   [7]  1.0821039  1.1386970  1.2252254  1.3641016  1.6505646  1.7966908
#>  [13]  1.8865232  1.8982560  2.3147437  2.3242932  2.3595847  2.3962857
#>  [19]  2.4142236  2.4164476  2.4515318  2.4559213  2.5088731  2.8756229
#>  [25]  2.9490862  3.0344741  3.1364854  3.1582254  3.2276375  3.2283513
#>  [31]  3.3529494  3.3870680  3.4811082  3.5791198  3.6594506  3.7066318
#>  [37]  3.7305269  3.7351765  3.7537738  3.7666453  3.8419602  3.8645643
#>  [43]  3.9331921  4.0612357  4.2052891  4.4113183  4.4591223  4.4856190
#>  [49]  4.7390957  4.8429651  4.8601339  4.8677843  4.9474937  4.9655662
#>  [55]  5.0192617  5.1075602  5.1605100  5.2572109  5.2702445  5.3719475
#>  [61]  5.4971349  5.4987108  5.5964307  5.6583688  5.7016005  5.7435824
#>  [67]  5.8141455  5.8987583  5.9243166  5.9276891  5.9472630  5.9486004
#>  [73]  6.0340074  6.1351050  6.1920048  6.3999147  6.6621396  6.8453858
#>  [79]  6.8587893  6.9439245  7.0017352  7.0541344  7.2311462  7.3652745
#>  [85]  7.3897437  7.6329481  7.6977100  7.8835715  7.9586442  8.0911328
#>  [91]  8.1754997  8.4353185  8.4476586  8.5003673  8.5091307  8.7800235
#>  [97]  8.8250077  8.9327908  8.9690600  9.0067103  9.1176893  9.1931784
#> [103]  9.4945033  9.5053692  9.6159107  9.6951914 10.0000000
#> 
#> attr(,"class")
#> [1] "pt"