Estimate probabilities to be in each state

Estimate probabilities to be in each state

estimate_pt(data, NAafterTmax = FALSE)

Arguments

data
data.frame containing id, id of the trajectory, time, time at which a change occurs and state, associated state.

NAafterTmax
if TRUE, return NA if t > Tmax otherwise return the state associated with Tmax (useful when individuals has different lengths)

Value

A list of two elements:

t: vector of time

pt: a matrix with K (= number of states) rows and with length(t) columns containing the probabilities to be in each state at each time.

See also

Author

Cristian Preda, Quentin Grimonprez

Examples

# Simulate the Jukes-Cantor model of nucleotide replacement 
K <- 4
PJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_PJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, P = PJK, lambda = lambda_PJK, Tmax = 10)

d_JK2 <- cut_data(d_JK, 10)

# estimate probabilities
estimate_pt(d_JK2)

#> $pt
#>   0 0.068 0.094 0.213 0.232 0.379 0.398 0.493 0.506 0.516 0.536 0.722 1.137
#> 1 1   0.9   0.8   0.9   0.8   0.7   0.6   0.6   0.5   0.4   0.4   0.4   0.3
#> 2 0   0.0   0.0   0.0   0.0   0.0   0.1   0.1   0.1   0.1   0.0   0.1   0.1
#> 3 0   0.1   0.1   0.0   0.1   0.2   0.2   0.3   0.3   0.3   0.4   0.3   0.4
#> 4 0   0.0   0.1   0.1   0.1   0.1   0.1   0.0   0.1   0.2   0.2   0.2   0.2
#>   1.203 1.267 1.298 1.331 1.393 1.491 1.498 1.513 1.572 1.671 1.688 1.713 1.744
#> 1   0.3   0.4   0.5   0.5   0.4   0.3   0.4   0.4   0.5   0.6   0.5   0.4   0.3
#> 2   0.0   0.0   0.0   0.1   0.1   0.1   0.0   0.0   0.0   0.0   0.0   0.1   0.1
#> 3   0.4   0.3   0.2   0.2   0.3   0.4   0.4   0.5   0.4   0.4   0.5   0.5   0.6
#> 4   0.3   0.3   0.3   0.2   0.2   0.2   0.2   0.1   0.1   0.0   0.0   0.0   0.0
#>   1.914 2.104 2.154 2.234 2.295 2.397 2.572 2.614 2.802 2.923 3.013 3.069 3.107
#> 1   0.2   0.3   0.2   0.2   0.2   0.2   0.2   0.1   0.1   0.1   0.2   0.3   0.4
#> 2   0.1   0.1   0.2   0.2   0.2   0.2   0.2   0.2   0.3   0.2   0.2   0.2   0.2
#> 3   0.7   0.6   0.6   0.5   0.4   0.3   0.2   0.2   0.2   0.2   0.2   0.2   0.1
#> 4   0.0   0.0   0.0   0.1   0.2   0.3   0.4   0.5   0.4   0.5   0.4   0.3   0.3
#>   3.215 3.216 3.358 3.387 3.393 3.59 3.645 3.679 3.718 3.783 3.935 3.944 3.961
#> 1   0.4   0.3   0.2   0.1   0.0  0.1   0.1   0.1   0.2   0.2   0.1   0.1   0.1
#> 2   0.2   0.2   0.2   0.3   0.3  0.3   0.3   0.3   0.2   0.3   0.4   0.5   0.5
#> 3   0.0   0.1   0.1   0.1   0.2  0.1   0.2   0.3   0.3   0.3   0.3   0.2   0.3
#> 4   0.4   0.4   0.5   0.5   0.5  0.5   0.4   0.3   0.3   0.2   0.2   0.2   0.1
#>   4.056 4.077 4.135 4.201 4.261 4.426 4.439 4.579 4.614 4.686 4.944 5.014 5.258
#> 1   0.1   0.2   0.1   0.1   0.2   0.2   0.2   0.2   0.3   0.4   0.4   0.3   0.2
#> 2   0.4   0.4   0.5   0.4   0.3   0.2   0.2   0.2   0.1   0.1   0.0   0.0   0.0
#> 3   0.3   0.2   0.2   0.2   0.2   0.2   0.3   0.2   0.2   0.1   0.2   0.3   0.4
#> 4   0.2   0.2   0.2   0.3   0.3   0.4   0.3   0.4   0.4   0.4   0.4   0.4   0.4
#>   5.371 5.404 5.49 5.492 5.496 5.803 5.867 5.968 6.155 6.188 6.364 6.495 6.731
#> 1   0.3   0.4  0.3   0.3   0.3   0.4   0.5   0.6   0.6   0.5   0.6   0.5   0.4
#> 2   0.0   0.0  0.1   0.0   0.1   0.1   0.1   0.1   0.2   0.2   0.2   0.3   0.3
#> 3   0.4   0.3  0.3   0.3   0.3   0.3   0.3   0.2   0.2   0.2   0.2   0.2   0.3
#> 4   0.3   0.3  0.3   0.4   0.3   0.2   0.1   0.1   0.0   0.1   0.0   0.0   0.0
#>   6.766 6.817 7.122 7.193 7.205 7.328 7.406 7.472 7.543 7.626 7.643 7.84 7.952
#> 1   0.4   0.3   0.3   0.3   0.3   0.3   0.3   0.4   0.5   0.6   0.5  0.4   0.3
#> 2   0.2   0.2   0.3   0.2   0.2   0.3   0.3   0.3   0.3   0.2   0.2  0.2   0.2
#> 3   0.3   0.3   0.3   0.4   0.3   0.3   0.2   0.1   0.1   0.1   0.1  0.1   0.2
#> 4   0.1   0.2   0.1   0.1   0.2   0.1   0.2   0.2   0.1   0.1   0.2  0.3   0.3
#>   8.012 8.146 8.279 8.342 8.685 8.729 8.818 8.874 9.051 9.312 9.341 9.345 9.419
#> 1   0.3   0.2   0.3   0.2   0.2   0.1   0.1   0.2   0.2   0.3   0.2   0.2   0.2
#> 2   0.1   0.1   0.1   0.2   0.2   0.2   0.2   0.2   0.2   0.2   0.2   0.2   0.1
#> 3   0.3   0.3   0.3   0.3   0.4   0.5   0.6   0.5   0.4   0.4   0.5   0.4   0.4
#> 4   0.3   0.4   0.3   0.3   0.2   0.2   0.1   0.1   0.2   0.1   0.1   0.2   0.3
#>   9.491 9.553 9.769 9.851 9.896 9.932  10
#> 1   0.2   0.1   0.1   0.2   0.1   0.2 0.2
#> 2   0.2   0.3   0.3   0.3   0.3   0.3 0.3
#> 3   0.4   0.4   0.3   0.3   0.4   0.3 0.3
#> 4   0.2   0.2   0.3   0.2   0.2   0.2 0.2
#> 
#> $t
#>   [1]  0.00000000  0.06812191  0.09407343  0.21293965  0.23243533  0.37865274
#>   [7]  0.39823875  0.49290512  0.50632344  0.51555204  0.53640001  0.72223547
#>  [13]  1.13729030  1.20340077  1.26661544  1.29843801  1.33087166  1.39268723
#>  [19]  1.49120494  1.49753739  1.51290936  1.57216313  1.67054559  1.68830956
#>  [25]  1.71314911  1.74404033  1.91410326  2.10439694  2.15370359  2.23437159
#>  [31]  2.29486908  2.39748246  2.57176157  2.61375961  2.80167885  2.92308796
#>  [37]  3.01314435  3.06939936  3.10660593  3.21517746  3.21607477  3.35821180
#>  [43]  3.38656210  3.39345804  3.58966190  3.64451122  3.67895971  3.71849528
#>  [49]  3.78277491  3.93482869  3.94352636  3.96125836  4.05630179  4.07699379
#>  [55]  4.13500358  4.20115296  4.26053865  4.42558331  4.43866916  4.57883853
#>  [61]  4.61434014  4.68564380  4.94364268  5.01384727  5.25771843  5.37098430
#>  [67]  5.40354617  5.48966555  5.49196344  5.49635825  5.80325655  5.86695829
#>  [73]  5.96827706  6.15522327  6.18811990  6.36364489  6.49492164  6.73086881
#>  [79]  6.76627531  6.81729238  7.12217461  7.19337980  7.20533849  7.32799933
#>  [85]  7.40604704  7.47219401  7.54255018  7.62557304  7.64342023  7.84049562
#>  [91]  7.95172952  8.01214965  8.14569902  8.27896104  8.34177786  8.68483849
#>  [97]  8.72902727  8.81803971  8.87432928  9.05117089  9.31185161  9.34096042
#> [103]  9.34509263  9.41919924  9.49061492  9.55277930  9.76940166  9.85093148
#> [109]  9.89625250  9.93224730 10.00000000
#> 
#> attr(,"class")
#> [1] "pt"